Added support for sprite-based events
/***** CODE SAMPLE *****/
sprite1.onClick("exit");
/***** END CODE SAMPLE *****/

Got over 1000 lines of code : 

[gmatheron@tin game]$ cat *.java|wc
   1208    3681   40375


git-svn-id: http://javascool.googlecode.com/svn@81 17e3714e-118d-3e7a-01ef-bee73ecf6174

+++ trunk/src/org/javascool/proglets/game/Rectangle.java
@@ -15,13 +15,13 @@ import org.javascool.tools.Macros;
 public class Rectangle extends Geometry implements Drawable {
     private boolean m_solid;
     
-    public Rectangle(float x, float y, float w, float h, boolean solid) {
+    public Rectangle(int x, int y, int w, int h, boolean solid) {
         super(x,y,w,h);
         m_solid=solid;
         ((Panel)Macros.getProgletPanel()).addItem(this);
     }
     
-    public Rectangle(float x, float y, float w, float h) {
+    public Rectangle(int x, int y, int w, int h) {
         super(x,y,w,h);
         m_solid=true;
         ((Panel)Macros.getProgletPanel()).addItem(this);
@@ -30,10 +30,10 @@ public class Rectangle extends Geometry implements Drawable {
     @Override
     public void draw(Graphics g) {
         if (m_solid) {
-            g.fillRect((int)getX(), (int)getY(), (int)getWidth(), (int)getHeight());
+            g.fillRect(getX(), getY(), getWidth(), getHeight());
         }
         else {
-            g.drawRect((int)getX(), (int)getY(), (int)getWidth(), (int)getHeight());
+            g.drawRect(getX(), getY(), getWidth(), getHeight());
         }
     }
     private static final Logger LOG = Logger.getLogger(Rectangle.class.getName());
